Fitur admin complete

+++ routes/web.php
@@ -2,6 +2,7 @@
 
 use Illuminate\Support\Facades\Route;
 use App\Http\Controllers\PTNController;
+use App\Http\Controllers\AdminController;
 
 /*
 |--------------------------------------------------------------------------
@@ -15,20 +16,18 @@ use App\Http\Controllers\PTNController;
 */
 
 Route::get('/', [PTNController::class, 'index']);
-Route::post('/', [PTNController::class, 'store']);
-Route::put('/', [PTNController::class, 'update']);
-Route::delete('/', [PTNController::class, 'destroy']);
+Route::post('/', [PTNController::class, 'store'])->middleware('auth');
 
 Route::get('/about', function () {
     return view('about');
 });
-Route::get('/admin', function () {
-    return view('admin');
-})->middleware('auth');
-Route::get('/login', function () {
-    return view('login');
-})->name('login');
+Route::get('/admin', [AdminController::class, 'index'])->middleware('auth');
+Route::post('/login', [AdminController::class, 'authenticate']);
+Route::get('/logout', [AdminController::class, 'logout']);
+Route::get('/login', [AdminController::class, 'login'])->name('login');
 
-Route::get('/create', [PTNController::class, 'create']);
-Route::get('/edit/{ptn}', [PTNController::class, 'edit']);
+Route::get('/create', [PTNController::class, 'create'])->middleware('auth');
+Route::get('/edit/{ptn}', [PTNController::class, 'edit'])->middleware('auth');
+Route::delete('/{ptn}', [PTNController::class, 'destroy'])->middleware('auth');
 Route::get('/{ptn}', [PTNController::class, 'show']);
+Route::put('/{ptn}', [PTNController::class, 'update'])->middleware('auth');
